+++ CHANGELOG.md
@@ -106,6 +106,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Refactor `OrderCancel` mutation for multiple warehouses - #5554 by @fowczarek
 - Add negative weight validation - #5564 by @fowczarek
 - Add error when user pass empty object as address - #5585 by @fowczarek
+- Fix payment creation without shipping method - #5444 by @d-wysocki
 - Fix checkout and order flow with variant without inventory tracking - #5599 by @fowczarek
 - Fixed JWT expired token being flagged as unhandled error rather than handled. - #5603 by @NyanKiyoshi
 - Refactor read-only middleware - #5602 by @maarcingebala